+++ HISTORY.md
@@ -1,4 +1,39 @@
 
+0.1.0alpha5 / 2012-10-05 
+==================
+
+  * froze deps
+
+0.1.0alpha4 / 2012-08-01 
+==================
+
+  * send last modified in header
+
+0.1.0alpha3 / 2012-07-31 
+==================
+
+  * ejs support for .html/.ejs files
+
+0.1.0alpha2 / 2012-07-26 
+==================
+
+  * fixed bin
+
+0.1.0alpha1 / 2012-07-26 
+==================
+
+  * re-wrote bin using optimist
+  * pass in yaml/json file for template data
+  * read from config file
+  * re-wrote preview, added unit tests
+  * tests for jade as input file
+  * more jade tests
+  * jade unit tests
+  * markdown unit tests, started refactor
+  * bump package
+  * Release 0.0.17
+  * added index.js
+
 0.0.17 / 2012-07-05 
 ==================